pyproject.toml
[project]
name = "BitMapGen"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "freetype-py>=2.5.1",
    "ipython>=8.18.1",
    "numpy>=2.0.2",
    "pillow>=11.1.0",
]

[project.scripts]
bitmapgen = "main:main"
